// Resturant.Internal.Web/ClientApp/src/app/+auth/models/user.ts
export class User {
  id?: string;
  email?: string;
  firstName?: string;
  lastName?: string;
  phoneNumber?: string;
  name?: string;
  imageUrl?: string;
  status?: string;
  membershipType?: string;
  institutionId?: string;
  institutionName?: string;
  emailConfirmed?: boolean;
  roles?: string[];
}